Use floor instead of floorf in the sine node

+++ src/libaudioverse/nodes/sine.cpp
@@ -50,7 +50,7 @@ void LavSineNode::process() {
 			phase+=phaseDelta;
 		}
 	}
-	phase -=floorf(phase);
+	phase -=floor(phase);
 }
 
 void LavSineNode::reset() {
